Move code to it's own encapsulation
